Float error

+++ models.py
@@ -5,7 +5,7 @@ class ResumeBullet(BaseModel):
     content: str = Field(..., description="Content of the resume bullet")
     feedback: str = Field(..., description="Feedback on the resume bullet")
     rewrites: list[str] = Field(default=[], max_items=2, description="Possible rewrites for the resume bullet")
-    score: int = Field(..., ge=0, le=10, description="Score for the resume bullet")
+    score: float = Field(..., ge=0, le=10, description="Score for the resume bullet")
 
 class ResumeExperience(BaseModel):
     bullets: list[ResumeBullet] = Field(..., description="List of resume bullets for the experience")
@@ -19,7 +19,7 @@ class ResumeProject(BaseModel):
 class FormattingAspect(BaseModel):
     issue: bool = Field(..., description="Whether there is an issue with this aspect")
     feedback: str = Field(..., description="Feedback on this aspect")
-    score: int = Field(..., ge=0, le=10, description="Score for this aspect")
+    score: float = Field(..., ge=0, le=10, description="Score for this aspect")
 
 class FormattingFeedback(BaseModel):
     font_consistency: FormattingAspect = Field(..., description="Assessment of font consistency")
@@ -36,7 +36,7 @@ class FormattingFeedback(BaseModel):
     page_utilization: FormattingAspect = Field(..., description="Assessment of efficient use of page space")
     is_single_page: FormattingAspect = Field(..., description="Assessment of whether the resume is single-page")
     consistency: FormattingAspect = Field(..., description="Assessment of consistency across all sections")
-    overall_score: int = Field(..., ge=0, le=10, description="Overall formatting score")
+    overall_score: float = Field(..., ge=0, le=10, description="Overall formatting score")
 
 class ResumeFeedback(BaseModel):
     experiences: list[ResumeExperience] = Field(..., description="List of experiences in the resume")
